backend/requirements.txt
aiohttp_cors==0.7.0
aiohttp==3.7.4.post0
python-socketio
python-dotenv==0.19.2
